Use the same types in `for` statements.
This removes lots of simple compiler warnings on 'signed' and 'unsigned'
comparisons (reported in #128). Please note that there are two types of
changes. It is because this does not change any function signatures.
  * In most cases: `int` --> `size_t`.
  * In `lookup_table_op.cc`, `string_to_number_op.cc`, and
    `sparse_tensor.h`: `size_t` --> int.

+++ tensorflow/core/kernels/lookup_table_op.cc
@@ -76,7 +76,7 @@ class HashTable : public InitializableLookupTable {
 
     const auto key_values = keys.flat<K>();
     const auto value_values = values.flat<V>();
-    for (size_t i = 0; i < key_values.size(); ++i) {
+    for (int i = 0; i < key_values.size(); ++i) {
       const K& key = key_values(i);
       const V& value = value_values(i);
       const V& previous_value = gtl::LookupOrInsert(table_.get(), key, value);
@@ -95,7 +95,7 @@ class HashTable : public InitializableLookupTable {
     const auto key_values = key.flat<K>();
     auto value_values = value->flat<V>();
 
-    for (size_t i = 0; i < key_values.size(); ++i) {
+    for (int i = 0; i < key_values.size(); ++i) {
       value_values(i) =
           gtl::FindWithDefault(*table_, key_values(i), default_val);
     }
